[REF] l10n_it_fatturapa_pec: Use SdI channel's send method
