+++ l10n_it_fatturapa_pec/__init__.py
@@ -1,2 +1,3 @@
+#  License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
 from . import models
-from . import wizard
