+++ l10n_it_fatturapa_pec/tests/test_e_invoice_response.py
@@ -30,7 +30,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         e-invoice to 'validated'"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, "IT03339130126_00009.xml")
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file(
             "POSTA CERTIFICATA_ Ricevuta di consegna 6782414.txt"
@@ -45,7 +45,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         """Receiving a 'CONSEGNA' posts a mail.message in the e-invoice"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, "IT03339130126_00009.xml")
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file("CONSEGNA_ IT03339130126_00009.xml.txt")
 
@@ -70,7 +70,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         """Receiving a 'ACCETTAZIONE' posts a mail.message in the e-invoice"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, "IT03339130126_00009.xml")
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file("ACCETTAZIONE_ IT03339130126_00009.xml.txt")
 
@@ -151,7 +151,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         self.env.company.vat = "IT14627831002"
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, "IT14627831002_02621.xml")
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file("POSTA CERTIFICATA_mancata_consegna.txt")
