--- l10n_it_fatturapa_pec/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import send_pec
